requirement.txt
fastapi==0.95.0
uvicorn==0.22.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
pydantic==1.10.7
azure-storage-blob==12.14.0